move `GET /contingency-lists` endpoint to supervision controller and update test cases
