+++ src/test/java/org/gridsuite/actions/server/ContingencyListControllerTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2020, RTE (http://www.rte-france.com)
+ * Copyright (c) 2026, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
@@ -401,7 +401,7 @@ class ContingencyListControllerTest {
     }
 
     private int getContingencyListsCount() throws Exception {
-        String res = mvc.perform(get("/" + VERSION + "/contingency-lists")
+        String res = mvc.perform(get("/" + VERSION + "/supervision/contingency-lists")
                         .contentType(APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andExpect(content().contentTypeCompatibleWith(APPLICATION_JSON))
